nexus-api/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
cryptography==41.0.3
boto3==1.28.44
psycopg2-binary==2.9.7
sqlalchemy==2.0.20